Vibeyf-ai/requirements.txt
# Dépendances pour l'API backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Dépendances existantes
sentence-transformers>=2.2.2
torch>=2.1.0
pandas>=2.1.3
numpy>=1.24.0
scikit-learn>=1.3.2
google-generativeai>=0.8.0
python-dotenv>=1.0.0
